SSH_HONEYPOT_IMPLEMENTER/requirements.txt
paramiko==2.11.0
mysql-connector-python==8.0.33
